Extract _update_cache.py from _legacy.py
